pyproject.toml


[project]
authors = [{ name = "James Hodgkinson", email = "james@terminaloutcomes.com" }]
requires-python = "~=3.12"
dependencies = [
    "meshtastic<3.0.0,>=2.3.8",
    "pydantic<3.0.0,>=2.7.1",
    "loguru<1.0.0,>=0.7.2",
    "click<9.0.0,>=8.1.7",
    "esptool<5.0.0,>=4.8.1",
    "pyyaml<7.0.0,>=6.0.1",
    "paho-mqtt<3.0.0,>=2.1.0",
]
name = "meshtastic-tools"
version = "0.1.0"
description = ""
readme = "README.md"

[project.scripts]
layer-configs = "meshtastic_tools.layer_configs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["collector"]

[tool.mypy]
plugins = "pydantic.mypy"

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "types-pyyaml>=6.0.12.20250516",
]
